Update patch set 2
Patch Set 2:

I'm not sure. I'd argue that the initial implementation was problematic because it was extending the styling of Toast to all notifications. If Minerva users wanted to see an error icon, they'd use the Toast provided here. I'd say that this is the correct behavior as we're only styling toasts and not mw.notifications.

To me, cleaner code (removed LESS code) outweighs the alternative, which is not showing the icon in some cases, where users bypass Minerva's Toast and use mw.notification directly. If they want to be Minerva compliant they should use Toast.

Patch-set: 2
Label: Verified=0