Update patch set 3
Patch Set 3:

>I'm not sure. I'd argue that the initial implementation was problematic because it was extending the styling of Toast to all notifications. 
but it was doing that intentionally.

>  If Minerva users wanted to see an error icon, they'd use the Toast provided here. I'd say that this is the correct behavior as we're only styling toasts and not mw.notifications.
We should at minimum document this in the commit message and Toast class. At the moment Toast styles both.

> To me, cleaner code (removed LESS code) outweighs the alternative, which is not showing the icon in some cases, where users bypass Minerva's Toast and use mw.notification directly. If they want to be Minerva compliant they should use Toast.

To be fair.. they are not bypassing the toast. The Minerva toast only works in Minerva. So from what you'd saying if an extension wants to work with Minerva and Vector they'd need to have 2 code paths and deal with asynchronous loading. That doesn't seem helpful to them and pushes the tech debt from us onto them. The mw-notification-error class will still be present and there will be toasts that are errors that are not styled as errors which may cause confusion to end users.

I'm sure Florian has some views on this since he wrote the original ticket so I've added him just in case but I really do think we should be converging more on mw.notify than our own toast.

I'm going to poke around the code in core today as I think it would be valuable to have mw.notify and Toast use identical code paths - purely because we'd no longer have to replicate it.

Patch-set: 3
Label: Verified=0