Ensure mDiffEngine is always available
The methods have been renamed to make it clearer what the purpose
of these methods are.

They are also made private to make it clear they should not be
used out of context.

addMobileDiff should only be called if InlineDifferenceEngine is
used as getWarningMessageText is not a method of DifferenceEngine

This is a follow up to
I68cf50f5dd339f34802d70df1f32d2c3390944a3

Bug: T167461
Change-Id: Ia59d76769960630859c9666acc5a28317e9272b7