Update patch set 1
Patch Set 1:

In what way are the hlist meant to render differently in Minerva from the default? It seems like they doing the same thing: Fold the list into a sequence of inline elements, with bullet separators between them.

The difference seems rather arbitrary:

* Instead of using a space before/after within the "content", Minerva hard-codes a px margin/padding. Arguably, the code approach scales better here, and possibly better for usability (assuming the margin won't be interpreted as a space when the content is copied, resulting in "Foo*Bar" instead of "Foo * Bar").
* There is something about a linkColor that I don't quite understand.

The main difference I do understand is that this version doesn't add support for older browsers. Which makes it an optimisation only. Is that the rationale?

If so, I can still support this, but t would be nice if the rest of the implementation is made consistent for maintainability. + a big FIXME comment saying to manually keep this in sync +  tech debt task.

Patch-set: 1