Update patch set 2
Patch Set 2:

> In what way are the hlist meant to render differently in Minerva from the default? It seems like they doing the same thing: Fold the list into a sequence of inline elements, with bullet separators between them.

Yep there's definitely some overlap. In particular there are lots of css rules in MediaWiki:Mobile.css on enwiki that I'd like to consolidate.

>The difference seems rather arbitrary:
Instead of using a space before/after within the "content", Minerva hard-codes a px margin/padding. Arguably, the code approach scales better here, and possibly better for usability (assuming the margin won't be interpreted as a space when the content is copied, resulting in "Foo*Bar" instead of "Foo * Bar").

it's a little arbitrary but I think it should be up to the skin to control. For instance a fictional pirate skin might want to put a skull and crossbones between them :)

> There is something about a linkColor that I don't quite understand.
Could you elaborate? The idea is to make the colour the same as the link.

> The main difference I do understand is that this version doesn't add support for older browsers. Which makes it an optimisation only. Is that the rationale?

Partly, but as I point out there are more subtle tweaks. These could arguably be done anyway using overrides but I think it's better to do either/or.

> If so, I can still support this, but t would be nice if the rest of the implementation is made consistent for maintainability. + a big FIXME comment saying to manually keep this in sync +  tech debt task.

That's my plan. I'd like to update the core module to have some defaults (expressed in stylesheets again) with the customisations/older browser support in skinStyles overridden by Minerva. Does this make sense?

Patch-set: 2
Label: Verified=0