Update patch set 1
Patch Set 1: Code-Review+1

I was curious if the tidy feature did anything useful, so I did SWAT a change to remove it e.g. turn it on. Monte's taking a look (apparently iOS do not use MCS). I say we hold off merging this until Wednesday as if it turns out its doing something useful maybe we'll want to keep it and make it default to on rather than remove.

If Monte notices nothing I think we can safely just throw this away.

Patch-set: 1
Label: Code-Review=+1