Fixups prior to Minerva/MobileFrontend separation
* Correct dependency tree - a MobileFrontend module should not
depend on a Minerva module - move "skins.minerva.icons.images.variants"
dependency to Minerva entry point
* Update jsduck config so that MainMenu is a global (MainMenu will
become an abstract concept and the code defined in Minerva)
* Messages are defined twice inside mobile.startup and skins.minerva.scripts
since the latter depends on the former, remove the messages from it
* skins.minerva.watchstar should not double define watchstar messages
that are already defined in mobile.watchstar module
* skins.minerva.talk module should not reuse a message key meant for another
purpose. Give it its own message.

Change-Id: I598e9230fc27aae46d8d55c63fb18d5e4aedf939