Update patch set 2
Patch Set 2: Code-Review-2

This patch is part of the review for the patch on Minerva against https://phabricator.wikimedia.org/T157995
I didn't want to step on Baha's toes but I think it is a valid approach to the solution. You're feedback is welcomed there.

Sorry for not making that clearer in the commit msg

Patch-set: 2
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2