Update patch set 2
Patch Set 2:

no, we do not know that. It might be related to the fact that our database is not consistent. This is the reason why there is logging introduced (it logs the url, and langObject) so we can find broken languages.

Patch-set: 2
Reviewer: Pmiazga <3945@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0