Update patch set 1
Patch Set 1:

This will not be mergeable unless the prettyify option is removed or the grunt job is updated and all other svgs.

The latter method if chosen will delay this a lot - I don't really have bandwidth right now to fix up the npm job and test that many icons.

Patch-set: 1