Update patch set 1
Patch Set 1:

> Also, is there anything that stops us from bumping the
 > MobileFrontend MediaWiki requirement to 1.30 and just assume
 > CommentStore is available?

Only if you want to maintain compatibility in MobileFrontend master with non-master versions of MediaWiki core. If you're fine with dropping that backwards compatibility, feel free to amend this patch to remove the unnecessary checks.

Since [[mw:Extension:MobileFrontend]] doesn't specify the (relatively new) 'compatibility policy' parameter to {{Extension}}, I erred on the side of caution when making this patch. Setting that parameter to 'rel' indicates that it's safe to not worry about compatibility with older versions of MediaWiki core when making changes like this.

Patch-set: 1