Third party main page logic is incorrect
On a wiki where special casing of main page is not enabled we should
not add the toggling indicator and JS as we do not run toggling code
on the main page. It is expected that main pages will be minimal content
that help users find their way around the site - they are not meant for
long form reading so section collapsing is not useful here.

When special casing is enabled we also do not run this code.

Changes:
* Call enableExpandableSections once and set it regardless of whether
main page special casing is enabled
* Mark setMainPage as deprecated - it is confusing and long term it will
be removed - it actually means "special case main page" - update documentation
to explain this.
* Logic for section collapsing being disabled in Formatter is removed. This should
now be a decision made by the caller

Bug: T176414
Change-Id: I20acfddede6ae013e3216e19eca280dd0a094691