Font changing is a "mobile" feature, not "minerva" feature
Since the font changing feature is managed via Special:MobileOptions
it should apply to all skins operating in mobile mode.

Additional changes:
Given it is only used in one place and not very reusable the only benefit
to having mobile.fontchanger in its own module is it is easier to unit test.
Given this is not happening, and there are better ways to do this
(e.g. see Popups extension), we merge the mobile.fontchanger module into
mobile.special.mobileoptions.scripts.fontchanger

Change-Id: I857cfe2d9be9fe49c04c860bc234384c787239b2