Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

This LGTM. The tests pass and the majority of the change is easily understood. I've left a question inline about the use of the MediaWikiLangTestCase parent class.

If other folk understand that part of the change better than me and want to +2 straight away, then I defer to your expertise :)

Also, thanks!

Patch-set: 3