Update patch set 2
Patch Set 2:

@Pmiazga 'server' uses $wgServer (potentially protocol-relative) and 'mobileserver' uses $wgCanonicalServer (always an absolute URL). That might be good or bad depending on how it's expected to be used. Browsers work better with relative URLs (absolute ones can result in mixed content warnings or very old browsers getting locked out), non-browser clients do not necessarily understand them. Wikimedia sites are HTTPS-only so this is as long as this is only going to be used by Wikimedia-specific code, absolute URLs are better.

When the mobile domain is the same, outputting the same domain still seems less surprising than not outputting anything, IMO.

The localhost thing is a Vagrant hack; $wgServer and $wgCanonicalServer are not supposed to use a different domain. Vagrant sets $wgServer dynamically from the current request to avoid redirects. IIRC the motivation was to allow setups scrips to run without relying on DNS resolution of *.local.wmftest.net which requires a network connection.

Patch-set: 2