Use ParserOutput stateless transforms when available
See MediaWiki core changes Ied5fe1a61 and I78b62ec3.

There's no need to feature-flag the extra parameter in calls to
ParserOutput::getText(), they'll just be ignored if this is merged
first.

This also changes the TOC-removing logic in MobileFrontendHooks to
actually remove the TOC from the HTML instead of relying on the stateful
ParserOutput::setTOCEnabled().

Change-Id: I3565868af824a08235ab5ce4a34145895ed0e74d