Use Html messageBox helpers
These helper methods are now upstreamed into core
(see T166915) and can be removed.

This was previously attempted but various calls to MobileUI
methods were accidentally retained. This attempt should hopefully
be more thorough.

Clarification:
* MessageBox.mustache is retained given its usage inside JavaScript
inside mobile.messageBox however it is updated to run as a hogan
template as now it is not needed for server side rendering. The
hasHeading hack is no longer necessary there.

Change-Id: Ib13800b9564dbd8d14a356d61823ca9810b511e1