Set tagline on Special:MobileOptions page
Restrict the existing setting of wikidata descriptions to pages
in the main namespace. This will allow us to make the Minerva skin
dumber about when it should render and when it should not.

Also set a tagline for the special page MobileOptions

Given, we anticipate other special pages to also have taglines,
a configuration variable MFSpecialPageTaglines is added.

Bug: T180095
Change-Id: I3f5886715f0632e67e46e077d362a988a648a123