Update patch set 3
Patch Set 3:

Here's a weird scenario:

Toggle the switch and refresh (not hard refresh) the page. You'll see the font selector (which means the beta mode is enabled), but the toggle switch is off.

Also, why are we saving the state in the back end if the toggle switch doesn't restore its state on another device?

Patch-set: 3
Label: Verified=0