Cookie must be set on server side
Setting the cookie on the client doesn't work.
It needs to be stored on the server.
This is because a call to MobileContext::setMobileMode must be
made to opt-in to beta, and cookies are not automatically forwarded from
the request to this method as needed. MobileOptions when submitted will
directly call this function.

We could of course update this behaviour must it would mean processing
cookies for every page view (not good!).

Submitting the form without an async save also avoids
the problem of having to refresh the page so after talking to Nirzar we
agreed to keep this simple and make it submit the form.

Changes:
* returnto parameter no longer honoured from Special:MobileOptions
* Remove MobileOptions::getURL which is dead code
* Successbox will only show in no-JavaScript mode. A toast is shown
in its place with JS.

Bug: T169369
Change-Id: I4028a0d37f51d8f6c6e64d49b52a76781b38676c