Update patch set 4
Patch Set 4: Code-Review+2

(10 comments)

This is definitely an improvement, but it's clear at some point in the future we should look at improving some of the documentation here.

I'm hoping that will happen naturally as and when we touch the code so I don't think it's necessary to do that now (unless you are interested in helping! :).

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2