Toggle the Beta toggle when user clicks on label
The beta checkbox is generated by PHP and then it's
"infused" via Javascript which lacks some fuctionalities.
To make it working we had to listen on toggle switch and
manually update the checkbox.

What we also have to do is
to listen on checkbox changes (triggered by clicking on the
label) and update the toggle plus submit the form. This task
does that.

Bug: T169807
Change-Id: I738d2dfb37608218c841a5d2014734faf5e16cae