Delay form submit by 0.25s
On iOS 11, mobile chrome 63.0 on iPhone Xs the button toggle gets stuck 
in middle as browser starts submiting the form. Browser tries to reload 
the page, and this cause all animations to stop. To quickly fix that let's
delay the form submission by 0.25seconds.
Value 250ms is taken from OOUI codebase, it's an animation time for
toggle button to switch from one state to another.
Tested on Chrome 61, Android 7.0.0

Bug: T169807
Change-Id: Ic06da614eef9f0bc1cf04705c2204bd5968ebda6