Remove PHPDoc block repeating the file name
I can see these have been added in a commit that aimed to improve the
PHPDoc comments in this code base back in 2014, see Ib5a9507. However,
I do not see the point in literally repeating the name of the file that
contains a comment. I assume this was to work around a bug in a specific
documentation generation tool.

Change-Id: I821845436d8a4cf62a95707d8662e205602c8ff0