Utilize the …::class feature instead of class names in strings
Yes, the …::class feature works, even if a class is not available. It's
technically still a string. However, most tools can much easier understand
that the string before …::class refers to a class name, and properly warn
when the class is renamed or moved to an other namespace.

note this conflicts with I4e47cdd! Whatever is merged first (and I believe
both can be merged), the other patch must be rebased then.

Change-Id: If8056135a212cf11017bce4c5fd08f83cd910d97