Hygiene: Begin refactoring of MobileFormatter mega class
After a discussion with Piotr we agreed that the MobileFormatter
code is becoming unmanageable and hard to test.

Going forward we will aim to reduce all transforms to simple classes
that take a DomElement and transform it into another DomElement.
This will improve testability and sanity.

As a first pass we will only update the lead paragraph transform
given that we are currently working on improving that feature in
T170006

To prove this is not specific to the lead paragraph we add the
AddMobileTocTransform to demonstrate this is extensible

Additional change:
* Move logic outside MobileFormatter to before creation of class
* Remove superfulous test - the lead paragraph transform is never
invoked outside main namespace so it's impossible to log things in
the main namespace. This test thus serves no purpose.

Bug: T170006
Change-Id: I1fb2dc9e7ad1fe0431819ce3ca6295b3df8410e0