Treat mw-stack elements as if they are infoboxes
If the first element of the page is an element with class mw-stack
that contains an infobox we should treat the entire element as if it's
an infobox and hoist the paragraph.

This overhauls a lot of the logic in MoveLeadParagraphTransformTest
to hopefully make it simpler. The logging of wrapped invalid infoboxes
now becomes dumb.

getInfoboxContainer is left generic to allow us to change the allowed
classes at a later date

Bug: T170006
Change-Id: I212504e9b1f9b055ab25c15645364a8a27ba6f91