Remove Special:MobileOptions/Languages
This was added back in 2012
In the old version of the language overlay when a language
did not exist we linked to this page. We don't do that anymore
since the new language overlay we launched in August 2016.

It's not linked to anywhere in the UI.

Should probably remove as it caused me a big surprise to not only
see it existed but to see I wrote the code haha

Change-Id: I9c7cca88c4f3df42f2ced93b72011f18d911acce