MobileFrontendFeaturesRegistration hook should be called earlier in stack
Most importantly hook should be called before the
RequestContextCreateSkinMobile hook as this is used to configure skins
so they behave differently in mobile mode. This hook is used by Minerva
for instance to enable categories and back to top beta features.

Currently, the features are not available to the skin at the time the
hook is called. This rectifies that.

Bug: T182362
Change-Id: I03839d840d02ed84a850baef1d8625ecdbe6970d