Hygiene: Remove MobileContext::getConfigVariable
Going forward we will make use of the FeatureManager for these kinds
of checks.

Given the complexity of the shouldShowWikibaseDescriptions method
I've not (yet) removed it, but have wired it up to the feature
management system.

We should also register features that have been promoted to stable
as if they are ever demoted to beta they will not appear in
Special:MobileOptions

Change-Id: I6aa1c66ec131a8db75d6e6128d4e3af78f351af0