Remove blacklisted content before applying transforms
Previously we removed all blacklisted elements AFTER applying
transforms such as moving lead paragraph.

This is unnecessary and can lead to problems, for example infoboxes
inside elements which are to be removed, might be moved above the
lead paragraph and skip the blacklisting altogether. This also leads
to unnecessary logging when the move is not possible.


HtmlFormatter::filterContent takes care of content removal so lets
do that BEFORE our transformations explicitly.

Bug: T185040
Change-Id: I18bc97fa37bdc53e61cb6de79fb6bc1158ffd9fe