Admins should not see surpressed content in feeds
In history and contributions pages surpressed content is currently
unhidden by default. To be consistent with desktop this should only
show when a query string parameter unhide is given.

To make admins happier we will now hide these by default. An admin
who needs to reveal them can use unhide=1 in query string. We do not
currently link to unhide in the UI, but an admin can click through to
the diff to get those details.

When testing check body, summary and user are correctly hidden and
unhidden for admins and users

Additional changes:
* DRY - much code is shared between Contributions and History so this
is upstreamed to MobileSpecialPageFeed

Bug: T118864
Change-Id: I016bf2a4e093d485ae6670bb1978ea5f70ab95d6