Update patch set 2
Patch Set 2:

> The patch fixes it because `button` default behavior used for search but `a` with a # will just keep the action on same page (top) and then performs the action.

Sounds like this should be mentioned in the commit message! It wasn't obvious to me at first that the button was inside a form element :)

> I've looked at the code to see if there is an event object for me to attach and override the default behavior and I've not seen so that's why I went for the `a` solution.
> I think it would make sense to use the preventDefault() if it's possible to have and event object to override.

Take a look at resources/mobile.special.nearby.scripts/nearby.js
refreshCurrentLocation
I think semantically and given I3efd07572c39 this makes more sense as a button so the preventDefault is probably a better way to go.

Patch-set: 2