Update patch set 12
Patch Set 12: Code-Review+2

This is fine.

The reason this is awkward is that the refresh icon being added to the form and replacing notifications is hacky (it doesn't show on desktop for example).

I'm fine with this solution but long term the solution would be:
https://phabricator.wikimedia.org/T125820

Patch-set: 12
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2