Fix talk page refresh after adding content
Talk pages were not being refreshed after adding content through the
"add discussion" button. The refresh is based on whether or not the
user was on a Talk page before opening the overlay.

This logic is determined by comparing the title of the previous page
to the current page. The `self.title` value had spaces but
`mw.config.get( 'wgPageName' )` had underscores, so the comparison was
always false if the page title, or namespace (like "User talk:") had a space.

Instead, adding `pageTitle` as an option and passing in the value from
`M.getCurrentPage().title`(which has spaces) in Minerva.

Bug: T184273
Change-Id: I83d68d149886a24bce519f3aaa3f16b8941309c9