Update patch set 1
Patch Set 1: Code-Review+1

Since this is an overlay, maybe this code relied on saved state so that if the underlying page changed (e.g., a user click a slow loading link) the overlay could detect that on save completion. I'm not sure how likely that is or whether a better fix would be to pass the correct title via the options object.

Patch-set: 1
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1