Do not unnecessarily bind Page to window load event
In T91882 we introduced a window load event on every Page render
to attempt to reveal the hash, however a later and similar bug
was raised in T141452 which correctly called the checkHash method.

Thus this code is not doing anything any more except making jQuery
migrate complain. Let's remove it.

Bug: T188346
Change-Id: I61ad7b363aacb9d6dffe70bfe62844dc015e187d