Update patch set 4
Patch Set 4: Code-Review-1

Jan caught a typo and an omission, otherwise LGTM.

If anything, it seems to me that having to reference `View.prototype.parseHTML` a few times says that maybe this should be a method in `util/html` or some other standalone module, which would be required when needed.

So consider that ^, but patch looks fine after fixing Jan's comments.

-1 for attention

Patch-set: 4
Label: Code-Review=-1