Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Seems fine, just left a question (same question is on a few files). If we call getWindow at the top of the file at the module level then when testing it would be a pain to mock. Should we use the function in functions as needed instead of at the module level? (some files do that in this patch, while others don't)

Patch-set: 2
Label: Code-Review=-1