Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Found something weird on search overlay.

Also, I have the same question as with the previous patch, so feel free to disregard or act on it depending on what you think on that one.

Should we call getDocument when needed instead of at the top of the modules to make testing easier (at some point?)?

Patch-set: 2
Label: Code-Review=-1