Introduce util.parseHTML and View.prototype.parseHTML
Always use local jQuery where possible.
To avoid side effects, jQuery find should run local to a
component, not globally. This means components will not be able
to impact other components, and we will be able to enforce this
in future by disabling usage of the jQuery global in eslint.

We will also explicitly use $.parseHTML rather than relying
on $( '<div>' ) to make it explicit where we are creating
new elements (in comparison to selecting elements)

The remaining usages relate to components accessing the window
object. These will be addressed in a follow up.

Bug: T173539
Change-Id: I464da5c528764feb769f3c1d232b172df810c62c