Don't store a cookie for mobileEditors any more
The mobileEditor cookie was added back in 2e0af840
for the purpose of a CentralNotice campaign. This campaign is long
over and there are plenty of other ways now to detect whether a
user has made mobile edits.

We shouldn't keep this around without good reason so lets remove it.

Additional changes:
* Drop unnecessary dependencies to jquery.cookie - this is not
used anywhere in the codebase

Change-Id: Ie60aca9450bc004dc63c38374cbb5f2051cb3e1c