Do not refresh routes on all hash changes
Note:
The /page/ and /coord routes are undocumented and undiscoverable.
We don't actively support them.

When visiting the Nearby page we need to load the nearby results.
These shouldn't refresh on an arbitary hash change e.g.
searching.

This will probably no longer be necessary when T125820
is completed as that will make the user click to see results.

Bug: T188784
Change-Id: Icbf2fe378ce7b4eaba97aa8da4dd9c8781ffd709