ContentProvider now routes all api queries via same API/adds lang links
While working on the language overlay, I was frustrated at not being able
to replicate production data.

This change to the ContentProvider routes all API queries made on the client
via the same content provider.

This is pretty powerful: it provides langlinks for the current page and allows
the editor and talk overlays to behave as they would in production (with the only
downside being due to a lack of login support saving will not work!)

Note for localhost development you'll need to make use of a Chrome extension
to ensure the API queries work others wise you'll get an exception in Chrome's
dev tools.

Change-Id: Iaa200eeb9f81eb31a8598d8983023373e715f17b