Right align RTL languages in language list
This right aligns any languages which are RTL in the language
overlay.

Given the existing limitations of API and core this uses a static
array of all known RTL languages. Although theoretically this could
go out of date, this is easy enough to maintain given the bugs go
reported.

To make clear this is a workaround, a warning is posted to the console
reminding us that this exists. It points to a bug number which allows us
to track and hopefully remove this code in future in favor of accessing
directly from the API or mw.config.

Bug: T66617
Change-Id: I84302265491404d7642793107ff0ed2a01092d13