Update patch set 1
Patch Set 1: Verified+1 Code-Review-1

This will increase first paint and speed index. We will need to get a sense of that before going ahead with this change.

It was purposely made non render blocking for this reason. The hope was templatestyles would solve this problem for editors.

Patch-set: 1
Label: Verified=+1
Label: Code-Review=-1