Update patch set 1
Patch Set 1:

I'm not sure we want submit inside the repo. We removed it purposely as it was causing confusion on the UI (people submitting without letting Jenkins test) and Jenkins is usually good at reporting bugs. Is it possible to only enable this on branches other than master?

Patch-set: 1