Update patch set 1
Patch Set 1:

This is the first part, we don't need the font-size magic IMO. We still can't rely on `contain` though (lib's default) as the widths of the icons is different due to 16px font-size. But setting font-size doesn't really make this better, paddings and widths should be adapting via media queries in the mid-term. 
Second part is changes in VE itself for the button paddings…

Patch-set: 1