Don't set wgScriptPath in content provider
Currently when enabled we rewrite all API requests via production.
The problem with this is because of CORS it requires a browser
extension to work. If users do not have one installed it breaks the
site. 

To allow us more testing workflows such as: 
https://phabricator.wikimedia.org/T189212#4104431 we can do this without
the content provider - just set wgScriptPath in LocalSettings.php.

Change-Id: If2301fba6ed78f2a4bfe8ae4e3f13d29b98776e7