Provide a more production-like environment for desktop
Given we are using reading web staging to test desktop we want
to have an environment closer to production. We don't run the Mobile
Formatter in desktop mode so shouldn't run it in our test environment
either. Actually what we want is to use the content provider
not the formatter and content provider.

Change-Id: Ia470057f1aecefcacc045be1ba23c3e3523f7cec