Update patch set 1
Patch Set 1: -Code-Review

> still failing in IE11 for me.

Scratch that. Fix does work on IE11.


I'm not 100% on this patch because it (understandably) doesn't address the root cause of this issue, which is buried somewhere deep in VE.


This patch ends up setting the scroll on the document body (or html) like, 100% of the time from my testing. This probably isn't the intended purpose of this `_resizeEditor` function, and like Piotr mentions, it has the downside of setting the cursor above the top menu bar sometimes, BUT, it also has an upside that this doesn't happen if you resize the window before typing:  https://pasteboard.co/HfcmMHs.png so that's good. 


I'm wondering though, if the solution sets the scroll on the document, whether the `_resizeEditor` function is even necessary, and if just setting the content area to 100% height will have the same effect.


(unless there are other use-cases where we don't want the editor to have 100% height).

Patch-set: 1