Update patch set 1
Patch Set 1:

As near as I can tell, the problem starts at the beginning. I'm guessing that the thinking was that we'd want to toss the 1st result for continuation but still wanted to request the maximum or at least a consistent number of remaining results. There's a note that we should strive to keep the PHP and JS implementations in sync but I don't think it's a requirement. An alternate implementation is to keep the same logic but drop the max to 49 so that adding one only goes 50 instead of 51. #AncientAstronauts

https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/MobileFrontend/+/66e00a9ee/javascripts/modules/watchlist/WatchListApi.js#56

Patch-set: 1
Reviewer: Niedzielski <2426@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0