Update patch set 4
Patch Set 4:

I'm not sure this is a problem.
The test failures I've been seeing recently are with
https://integration.wikimedia.org/ci/job/mwext-mw-selenium-jessie/8769/
(	Switch between mobile and desktop views: Switching from mobile view to desktop view.mp4	)
If you view this video it doesn't seem to be waiting long enough between steps which is a timeout issue... It seems to click the toggle link but not wait for the new view to load.

I also seem to have lost Job/Create permission so cannot check this on beta cluster making this  change risky until that happens:
https://www.mediawiki.org/wiki/Reading/Web/QA#Simulating_browser_test_run_on_an_unmerged_patchset.

Patch-set: 4