Update patch set 4
Patch Set 4: Code-Review+1

Hey guys, so I think you're both correct. 


This error, visible on this log 
https://integration.wikimedia.org/ci/job/mwext-mw-selenium-jessie/8769/console 
was caused by a timeout issue on the "Then I should see the desktop view" step, and not by this undefined step.  


This patch is still correct though, because if you look at previous "successful" logs, https://integration.wikimedia.org/ci/job/mwext-mw-selenium-jessie/8766/console you'll see that the "Then I should see the desktop view" step passes, but then the step "Given I am on the desktop site" after it is undefined, and the next three steps are just skipped... and the logger doesn't throw an error on undefined or skipped steps, which is basically useless... but that's another story. 

@Jdlrobson I believe the current tests pass **because** of this patch. If you look at the selenium log for this patch https://integration.wikimedia.org/ci/job/mwext-mw-selenium-jessie/8791/console you'll see the tests pass and there are no undefined steps anymore! nothing written in yellow! hurra!

Given that outcome, I'd say there are no risks to this patch, and I'm +1'ing. I think the only issue might be setting the cookie to a boolean, but in the end all cookies get converted to strings, and true.ToString() is still "true".

Patch-set: 4
Reviewer: Jdrewniak <2950@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1