Update Drawer appearance to bring better focus and spacing to drawers
The Drawer does not focus the user on the drawer enough. To better focus the
user, the existing content should fade mostly to black while the drawer is
open. The default cancel button is now given a default amount of space. The
collapseIcon is now a default of the Drawer and can be used in an
implementation's template.

Bug: T165535
Change-Id: I926918d0bd7f2176f188a2154dc5e99f6a8a7ad1