Fix parsing/i18n of no-results message
The mobile-frontend-search-content-no-results message used another message as an embedded piece
using {{int:}}, marking it bold. However, the JavaScript message framework does not seem to work
that way. The embedded message was being parsed in the default context, so the piece was always
in English ("lorem ipsum '''Search within pages''' dolor sit").

Fixed, so that it uses client-side JS parsing using mw.message().parse(), converting the wiki
markup to HTML <b> elements so that the formatting works.

Bug: T199156
Change-Id: I86da55bbfd77761849685e3150317e13a6a2fc0b