Update patch set 1
Patch Set 1:

I'm confused. Is there a bug open for this?

This is working perfectly fine as far as I'm aware:
https://imgur.com/a/x5qgRh3
but agree the use of `int` inside the message is confusing.

I believe this was intended to workaround loading  the parsingJavaScript on the client (which this patch introduces)

Can we open a phabricator ticket and associate it with this patch, please so I can understand the problem we're trying to address ? Thanks!

Patch-set: 1