Verify that data['pageprops'] exists before using it
We see lots of errors:
 - Notice: Undefined index: pageprops ApiMobileView.php on line 149
 - Warning: Invalid operand type was used: array_key_exists expects
an array or an object; false returned.

This patch should get rid of those errors. Before doing any
`addDescriptionToResult()` logic first verify that
`$data['pageprops']` exists and it is a valid array. There are
no other changes to logic, if pagepros are not available just skip
calling `addDescriptionToResult()`.

Additionally added typing of $pageprops argument.

Bug: T195872
Change-Id: I4ece5a01425a2a105f2f84fe116db35b2130e842