Simplify how table of contents is stripped from mobile view
There is a lot of indirection in our code. Via parser hook, the table
of contents is removed from input. We also remove HTML inside the MobileFormatter.
Rather than continuing to do this, it makes sense for this transform to live
in the transforms folder  inside AddMobileTocTransform

This makes the ContentProvider class compatible with the table of contents
(it will now show on API pulled content) and also makes it clearer where
the table of contents is stripped.

Changes:
* Introduce test for AddMobileTocTransformTest
* Remove outdated test from MobileFormatterTest

Bug: T193517
Change-Id: I1177e5e9913773af5e031cc3702d920faaf4aa56