OverlayManager should restore user's scroll when closing an overlay
Each Overlay records the current scrollTop position when it is shown and
restores it when a call to hide is made.
If a second overlay is shown then the second overlay causes the first
overlay to be hidden, thus when shown will always record the current
scrollTop as 0.
The OverlayManager is better equipped to deal with this case, so
should do similar checks. The code in Overlay's is not completely
redundant as certain overlays are not managed by the OverlayManager.

Additional changes:
* Stop using jQuery for scrollTop (see T188708 )

Bug: T189225
Change-Id: Icae22902eded907f04c73af69bb2751f6e57b983