Diff title should link to revision
There is currently no way in the mobile experience to access
a page at a given revision. On desktop the diff is shown alongside
the latest revision.

When viewing diffs it's important to be able to see the whole page
at the time of that revision.

The title should link to the id relating to that revision, not the
page as it currently is.

Change-Id: I95fb90784fe3a839cbe321e0e6fcfce121f359d6