Hygiene: Fix "@param $…" definitions
Various definitions in tests are missing types.

This patch adds a couple types to an existing block and removes another
block which is trickier to fix. The usage of param is questionable in
tests which are accompanied by a provider and we might consider
removing more of these in future.

Change-Id: I2b1f4f8797abd0b3680d00d670a90347554cceef