Update patch set 2
Patch Set 2:

Won't this always show the "this is an old revision" box above the page, because it always links to the oldid, even if it is the current revision ?

That might be a bit overkill. I can be mistaken, i haven't been able to test.

Patch-set: 2
Reviewer: TheDJ <61@e9e9afe9-4712-486d-8885-f54b72dd1951>