Headings can now be removed
Since the introduction of I0ed0255280f525d7726d2e09ceab15a76e2d6026
it is possible for the heading inside the table of contents to be
removed during the execution of makeSections inside tocTransform

When querying nodeName of an element that no longer exists a warning
is thrown.

We only use the firstHeading to work out which heading to make sections
with so we can cache this value earlier and avoiding this unnecessary
lookup.

Bug: T193517
Change-Id: I13c7b51d216db28bc389c7580d30ac63edf6bcdb